feat: Update dagger.json and add utilities for running commands
The changes in this commit include:

1. Added a new module `templating.rs` that provides a function to process template content for a new Dagger module.
2. Added a new module `command_utils.rs` that provides utility functions for running commands with output and formatting Go code.
3. Updated the `main.rs` file to use the new modules and remove redundant code.
4. Removed the `update_dagger_json` function from `main.rs` as it has been moved to a new module.
5. Added new functions `update_dagger_json`, `update_examples_dagger_json`, and `update_tests_dagger_json` in the `dagger_json` module to handle updating the `dagger.json` file for the main module, examples, and tests respectively.

These changes are focused on improving the overall structure and organization of the codebase, as well as introducing new utilities to make the development and maintenance of Dagger modules more efficient.
